+++ test/src/test_program.cpp
@@ -26,7 +26,10 @@ TEST(test_program, void_command)
 
     p.add_command("p", "print something", print_function);
     test_in.str("p\n p\n p\n q\n");
-    p.run(false, false);
+
+    p.set_print_help_on_run(false);
+    p.set_ready_indicator("");
+    p.run();
     EXPECT_EQ(3U, m_print_function.calls());
     EXPECT_EQ("", test_out.str());
 }
@@ -42,7 +45,10 @@ TEST(test_program, int_set_command)
     p.add_command<int>("a", "help", function, skipper::set<int>({0,7,42,1}));
 
     test_in.str("a\n 42\n a\n -10\n a\n 0\n a\n 7 a\n 11\n");
-    p.run(false, false);
+
+    p.set_print_help_on_run(false);
+    p.set_ready_indicator("");
+    p.run();
     EXPECT_EQ(3U, m_function.calls());
     EXPECT_TRUE(m_function.expect_calls().with(42).with(0).with(7).to_bool());
 }
@@ -59,7 +65,10 @@ TEST(test_program, float_range_command)
         skipper::range<float>(-4.0F, 7.2F));
 
     test_in.str("a\n 42\n a\n -1\n a\n 0\n a\n 7\n a\n 11\n");
-    p.run(false, false);
+
+    p.set_print_help_on_run(false);
+    p.set_ready_indicator("");
+    p.run();
     EXPECT_EQ(3U, m_function.calls());
     EXPECT_TRUE(m_function.expect_calls().with(-1).with(0).with(7).to_bool());
 }
@@ -76,6 +85,8 @@ TEST(test_program, trigger_errors)
 
     test_in.str("wrong key\n a\n wrong input\n q\n");
 
-    p.run(false, false);
+    p.set_print_help_on_run(false);
+    p.set_ready_indicator("");
+    p.run();
     EXPECT_NE("", test_out.str());
 }